Refactor SWE Agent code
